+++ backend/benefit4j-starter/src/main/resources/static/assets/benefitPlatformItem-BAcU0gNw.js
@@ -1,1 +1,0 @@
-import{f as e}from"./index-Beg_AXgH.js";const f=t=>e.get("benefit/api/v1/platform/items",{params:t});export{f as g};