+++ backend/benefit4j-starter/src/main/resources/static/assets/benefitPlatformItem-DSq0sjrq.js
@@ -1,0 +1,1 @@
+import{f as e}from"./index-h10vqzuC.js";const f=t=>e.get("benefit/api/v1/platform/items",{params:t});export{f as g};